requirements.txt
fastapi[all]>=0.100.0
uvicorn>=0.23.0
websockets>=11.0.3
pydantic>=2.0.0
python-dotenv>=1.0.0
twilio>=8.0.0
google-genai>=0.2.0
colorama>=0.4.6
requests>=2.31.0
SQLAlchemy>=2.0.0
httpx>=0.24.0
pg8000>=1.31.0
cloud-sql-python-connector[pg8000]>=1.9.0
